+++ src/lib/command.rs
@@ -205,7 +205,7 @@ pub(crate) fn run(step: &Step) {
         Some(ref command_string) => {
             let command_spec = match step.config.toolchain {
                 Some(ref toolchain) => {
-                    toolchain::wrap_command(&toolchain, &command_string, &step.config.args)
+                    toolchain::wrap_command(&toolchain, &command_string, &step.config.args, true)
                 }
                 None => CommandSpec {
                     command: command_string.to_string(),