+++ src/lib/toolchain_test.rs
@@ -1,16 +1,17 @@
 use super::*;
+use crate::test;
 use envmnt;
 
 #[test]
 #[should_panic]
 fn wrap_command_invalid_toolchain() {
-    wrap_command("invalid-chain", "true", &None);
+    wrap_command("invalid-chain", "true", &None, true);
 }
 
 #[test]
 fn wrap_command_none_args() {
     let channel = envmnt::get_or_panic("CARGO_MAKE_RUST_CHANNEL");
-    let output = wrap_command(&channel, "true", &None);
+    let output = wrap_command(&channel, "true", &None, test::is_not_rust_stable());
 
     assert_eq!(output.command, "rustup".to_string());
 
@@ -24,7 +25,7 @@ fn wrap_command_none_args() {
 #[test]
 fn wrap_command_empty_args() {
     let channel = envmnt::get_or_panic("CARGO_MAKE_RUST_CHANNEL");
-    let output = wrap_command(&channel, "true", &Some(vec![]));
+    let output = wrap_command(&channel, "true", &Some(vec![]), test::is_not_rust_stable());
 
     assert_eq!(output.command, "rustup".to_string());
 
@@ -42,6 +43,7 @@ fn wrap_command_with_args() {
         &channel,
         "true",
         &Some(vec!["echo".to_string(), "test".to_string()]),
+        test::is_not_rust_stable(),
     );
 
     assert_eq!(output.command, "rustup".to_string());